Split main.js into modules
