+++ index.html
@@ -165,6 +165,6 @@
     </section>
   </template>
 
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
